Move types in branded-types to src/types
