+++ src/routes/LoginWithOAuthApp.tsx
@@ -10,6 +10,7 @@ import { useNavigate } from 'react-router-dom';
 import { Button } from '../components/fields/Button';
 import { FieldInput } from '../components/fields/FieldInput';
 import { AppContext } from '../context/App';
+import type { ClientID, ClientSecret, Hostname, Token } from '../types';
 import type { LoginOAuthAppOptions } from '../utils/auth/types';
 import {
   getNewOAuthAppURL,
@@ -17,12 +18,6 @@ import {
   isValidHostname,
   isValidToken,
 } from '../utils/auth/utils';
-import type {
-  ClientID,
-  ClientSecret,
-  Hostname,
-  Token,
-} from '../utils/branded-types';
 import Constants from '../utils/constants';
 
 interface IValues {
