--- src/utils/branded-types.ts
@@ -1,15 +1,0 @@
-declare const __brand: unique symbol;
-
-type Brand<B> = { [__brand]: B };
-
-export type Branded<T, B> = T & Brand<B>;
-
-export type AuthCode = Branded<string, 'AuthCode'>;
-
-export type Token = Branded<string, 'Token'>;
-
-export type ClientID = Branded<string, 'ClientID'>;
-
-export type ClientSecret = Branded<string, 'ClientSecret'>;
-
-export type Hostname = Branded<string, 'Hostname'>;
