+++ src/utils/links.test.ts
@@ -1,9 +1,9 @@
 import { partialMockUser } from '../__mocks__/partial-mocks';
 import { mockGitHubCloudAccount } from '../__mocks__/state-mocks';
+import type { Hostname } from '../types';
 import type { Repository } from '../typesGitHub';
 import { mockSingleNotification } from './api/__mocks__/response-mocks';
 import * as authUtils from './auth/utils';
-import type { Hostname } from './branded-types';
 import * as comms from './comms';
 import Constants from './constants';
 import * as helpers from './helpers';
