+++ src/utils/storage.test.ts
@@ -1,5 +1,5 @@
 import { mockSettings } from '../__mocks__/state-mocks';
-import type { Token } from './branded-types';
+import type { Token } from '../types';
 import Constants from './constants';
 import { clearState, loadState, saveState } from './storage';
